Moved most AI functions into BaseAi.cs
