+++ Assets/Prefabs/Characters/QuestionMark.prefab
@@ -124,6 +124,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: a33cbad13bc72984aa27c9c5ec84e05c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  rotationSpeed: 90
 --- !u!1 &7427544645360850057
 GameObject:
   m_ObjectHideFlags: 0
