+++ Assets/Prefabs/Props/Pine_Tree.prefab
@@ -68,7 +68,7 @@ MeshRenderer:
   m_RenderingLayerMask: 1
   m_RendererPriority: 0
   m_Materials:
-  - {fileID: -3086382296381693183}
+  - {fileID: 2100000, guid: 4669023dba9b36242a67e8bac8b48e71, type: 2}
   m_StaticBatchInfo:
     firstSubMesh: 0
     subMeshCount: 0
@@ -109,15 +109,15 @@ Material:
     serializedVersion: 3
     m_TexEnvs:
     - _BumpSpecMap:
-        m_Texture: {fileID: 2800000, guid: bb5af8c3d0dcb584bb53eba9d27b8d8d, type: 3}
+        m_Texture: {fileID: 2800000, guid: 6f8cd5c3044e38a49abba98f88420635, type: 3}
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     - _MainTex:
-        m_Texture: {fileID: 2800000, guid: 6571b782484a4d740a1f027a9bcd5cd3, type: 3}
+        m_Texture: {fileID: 2800000, guid: 2d93c139a15fd8b47b24d5f4babfe982, type: 3}
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     - _TranslucencyMap:
-        m_Texture: {fileID: 2800000, guid: aaf8f8df1d038904f96562961cbb5364, type: 3}
+        m_Texture: {fileID: 2800000, guid: 9fdcc8a2ee5df494292809af6138cadf, type: 3}
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     m_Floats:
@@ -1065,19 +1065,19 @@ Material:
     serializedVersion: 3
     m_TexEnvs:
     - _BumpSpecMap:
-        m_Texture: {fileID: 2800000, guid: bb5af8c3d0dcb584bb53eba9d27b8d8d, type: 3}
+        m_Texture: {fileID: 2800000, guid: 6f8cd5c3044e38a49abba98f88420635, type: 3}
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     - _MainTex:
-        m_Texture: {fileID: 2800000, guid: 6571b782484a4d740a1f027a9bcd5cd3, type: 3}
+        m_Texture: {fileID: 2800000, guid: 2d93c139a15fd8b47b24d5f4babfe982, type: 3}
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     - _ShadowTex:
-        m_Texture: {fileID: 2800000, guid: fd226743f16d8914fa3c8e814c51ceb6, type: 3}
+        m_Texture: {fileID: 2800000, guid: 353bd299521fb13418ff9786cbe73227, type: 3}
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     - _TranslucencyMap:
-        m_Texture: {fileID: 2800000, guid: aaf8f8df1d038904f96562961cbb5364, type: 3}
+        m_Texture: {fileID: 2800000, guid: 9fdcc8a2ee5df494292809af6138cadf, type: 3}
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     m_Floats:
