+++ Assets/Prefabs/Props/Pine_Tree_Textures.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b0143cc7339df2049a044cfb4308d6a5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
