+++ Assets/Scripts/NPCs/PureChaser.cs
@@ -1,105 +1,95 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
+﻿using UnityEngine;
 using UnityEngine.AI;
 
-public class PureChaser : MonoBehaviour
+namespace NPCs
 {
-    public Transform targetTransform;
-    public float startChaseRadius = 100f;
-
-    private ChildRootMotionController rootMotionController;
-    private NavMeshAgent agent;
-    private bool shouldChase = false;
-
-    public const float reportReachedDistance = 2f;
-
-    public event Chaser.CollideWithPlayerAction OnCollideWithPlayer;
-    public delegate void ReachedDestinationAction();
-    public event ReachedDestinationAction OnReachDestination;
-
-    private void Awake()
+    public class PureChaser : BaseAi
     {
-        agent = Utils.GetRequiredComponent<NavMeshAgent>(this);
-        rootMotionController = Utils.GetRequiredComponentInChildren<ChildRootMotionController>(this);
-    }
+        public float startChaseRadius = 100f;
 
-    public void SetSpeed(float speed)
-    {
-        agent.speed = speed;
-    }
+        private ChildRootMotionController rootMotionController;
+        private bool shouldChase = false;
 
-    public void SetAnimationSpeed(float speed)
-    {
-        rootMotionController.SetAnimationSpeed(speed);
-    }
+        public const float ReportReachedDistance = 2f;
 
-    public void SetDestination(Vector3 position)
-    {
-        agent.SetDestination(position);
-    }
+        public delegate void ReachedDestinationAction();
+        public event ReachedDestinationAction OnReachDestination;
 
-    private void ChaseTarget()
-    {
-        if (shouldChase && targetTransform)
+        private void Awake()
         {
-            SetDestination(targetTransform.position);
-        }
-    }
-
-    private void Update()
-    {
-        if (!agent.isOnNavMesh)
-        {
-            SetMoving(false);
-            return;
+            agent = Utils.GetRequiredComponent<NavMeshAgent>(this);
+            rootMotionController = Utils.GetRequiredComponentInChildren<ChildRootMotionController>(this);
         }
 
-        CheckRemainingDistance();
-
-        // below are all about using targetTransform
-        if (!targetTransform || agent.speed <= 0f)
+        public void SetSpeed(float speed)
         {
-            SetMoving(false);
-            return;
-        }
-        else if (shouldChase)
-        {
-            SetMoving(true);
+            agent.speed = speed;
         }
 
-        if (!shouldChase && Vector3.Distance(transform.position, targetTransform.position) <= startChaseRadius)
+        public void SetAnimationSpeed(float speed)
         {
-            shouldChase = true;
+            rootMotionController.SetAnimationSpeed(speed);
         }
 
-        ChaseTarget();
-    }
-
-    private void SetMoving(bool isMoving)
-    {
-        rootMotionController.SetBool(Constants.ANIMATION_STEVE_MOVING, isMoving);
-    }
-
-    private void CheckRemainingDistance()
-    {
-        if (Vector3.Distance(transform.position, agent.destination) <= reportReachedDistance)
+        public void SetDestination(Vector3 position)
         {
-            OnReachDestination?.Invoke();
+            agent.SetDestination(position);
         }
-    }
 
-    private void OnCollisionEnter(Collision collision)
-    {
-        if (collision.gameObject.CompareTag(Constants.TAG_PLAYER))
+        private new void ChaseTarget()
         {
-            OnCollideWithPlayer?.Invoke();
+            if (shouldChase && targetTransform)
+            {
+                SetDestination(targetTransform.position);
+            }
         }
-    }
 
-    private void OnDrawGizmos()
-    {
-        Gizmos.color = Color.white;
-        Gizmos.DrawWireSphere(transform.position, startChaseRadius);
+        protected override void Update()
+        {
+            if (!agent.isOnNavMesh)
+            {
+                SetMoving(false);
+                return;
+            }
+
+            CheckRemainingDistance();
+
+            // below are all about using targetTransform
+            if (!targetTransform || agent.speed <= 0f)
+            {
+                SetMoving(false);
+                return;
+            }
+            else if (shouldChase)
+            {
+                SetMoving(true);
+            }
+
+            if (!shouldChase && Vector3.Distance(transform.position, targetTransform.position) <= startChaseRadius)
+            {
+                shouldChase = true;
+            }
+
+            ChaseTarget();
+        }
+
+        private void SetMoving(bool isMoving)
+        {
+            rootMotionController.SetBool(Constants.ANIMATION_STEVE_MOVING, isMoving);
+        }
+
+        private void CheckRemainingDistance()
+        {
+            if (Vector3.Distance(transform.position, agent.destination) <= ReportReachedDistance)
+            {
+                OnReachDestination?.Invoke();
+            }
+        }
+
+        private void OnDrawGizmos()
+        {
+            Gizmos.color = Color.white;
+            Gizmos.DrawWireSphere(transform.position, startChaseRadius);
+        }
     }
 }
